+++ modules/AWSPowerShell/Cmdlets/AmazonElasticsearch/Basic/Get-ESDomainList-Cmdlet.cs
@@ -1,0 +1,123 @@
+/*******************************************************************************
+ *  Copyright 2012-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.Elasticsearch;
+using Amazon.Elasticsearch.Model;
+
+namespace Amazon.PowerShell.Cmdlets.ES
+{
+    /// <summary>
+    /// Returns domain configuration information about the specified Elasticsearch domains,
+    /// including the domain ID, domain endpoint, and domain ARN.
+    /// </summary>
+    [Cmdlet("Get", "ESDomainList")]
+    [OutputType("Amazon.Elasticsearch.Model.ElasticsearchDomainStatus")]
+    [AWSCmdlet("Invokes the DescribeElasticsearchDomains operation against Amazon Elasticsearch.", Operation = new[] {"DescribeElasticsearchDomains"})]
+    [AWSCmdletOutput("Amazon.Elasticsearch.Model.ElasticsearchDomainStatus",
+        "This cmdlet returns a collection of ElasticsearchDomainStatus objects.",
+        "The service call response (type DescribeElasticsearchDomainsResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public class GetESDomainListCmdlet : AmazonElasticsearchClientCmdlet, IExecutor
+    {
+        /// <summary>
+        /// <para>
+        /// <para>The Elasticsearch domains for which you want information.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]
+        [Alias("DomainNames")]
+        public System.String[] DomainName { get; set; }
+        
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var context = new CmdletContext
+            {
+                Region = this.Region,
+                Credentials = this.CurrentCredentials
+            };
+            
+            if (this.DomainName != null)
+            {
+                context.DomainNames = new List<String>(this.DomainName);
+            }
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new DescribeElasticsearchDomainsRequest();
+            
+            if (cmdletContext.DomainNames != null)
+            {
+                request.DomainNames = cmdletContext.DomainNames;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(context.Credentials, context.Region);
+            try
+            {
+                var response = client.DescribeElasticsearchDomains(request);
+                Dictionary<string, object> notes = null;
+                object pipelineOutput = response.DomainStatusList;
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response,
+                    Notes = notes
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        
+        internal class CmdletContext : ExecutorContext
+        {
+            public List<String> DomainNames { get; set; }
+        }
+        
+    }
+}