+++ modules/AWSPowerShell/Cmdlets/AmazonElasticsearch/Basic/Remove-ESDomain-Cmdlet.cs
@@ -1,0 +1,133 @@
+/*******************************************************************************
+ *  Copyright 2012-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.Elasticsearch;
+using Amazon.Elasticsearch.Model;
+
+namespace Amazon.PowerShell.Cmdlets.ES
+{
+    /// <summary>
+    /// Permanently deletes the specified Elasticsearch domain and all of its data. Once a
+    /// domain is deleted, it cannot be recovered.
+    /// </summary>
+    [Cmdlet("Remove", "ESDomain", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.High)]
+    [OutputType("Amazon.Elasticsearch.Model.ElasticsearchDomainStatus")]
+    [AWSCmdlet("Invokes the DeleteElasticsearchDomain operation against Amazon Elasticsearch.", Operation = new[] {"DeleteElasticsearchDomain"})]
+    [AWSCmdletOutput("Amazon.Elasticsearch.Model.ElasticsearchDomainStatus",
+        "This cmdlet returns a ElasticsearchDomainStatus object.",
+        "The service call response (type DeleteElasticsearchDomainResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public class RemoveESDomainCmdlet : AmazonElasticsearchClientCmdlet, IExecutor
+    {
+        /// <summary>
+        /// <para>
+        /// <para>The name of the Elasticsearch domain that you want to permanently delete.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]
+        public String DomainName { get; set; }
+        
+        /// <summary>
+        /// This parameter overrides confirmation prompts to force 
+        /// the cmdlet to continue its operation. This parameter should always
+        /// be used with caution.
+        /// </summary>
+        [System.Management.Automation.Parameter]
+        public SwitchParameter Force { get; set; }
+        
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var resourceIdentifiersText = FormatParameterValuesForConfirmationMsg("DomainName", MyInvocation.BoundParameters);
+            if (!ConfirmShouldProceed(this.Force.IsPresent, resourceIdentifiersText, "Remove-ESDomain (DeleteElasticsearchDomain)"))
+            {
+                return;
+            }
+            
+            var context = new CmdletContext
+            {
+                Region = this.Region,
+                Credentials = this.CurrentCredentials
+            };
+            
+            context.DomainName = this.DomainName;
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new DeleteElasticsearchDomainRequest();
+            
+            if (cmdletContext.DomainName != null)
+            {
+                request.DomainName = cmdletContext.DomainName;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(context.Credentials, context.Region);
+            try
+            {
+                var response = client.DeleteElasticsearchDomain(request);
+                Dictionary<string, object> notes = null;
+                object pipelineOutput = response.DomainStatus;
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response,
+                    Notes = notes
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        
+        internal class CmdletContext : ExecutorContext
+        {
+            public String DomainName { get; set; }
+        }
+        
+    }
+}